Mapping + Prox sensors files updates + bugs fixes

+++ Sources/modules/headers/structs.h
@@ -13,8 +13,8 @@
 
 
 typedef struct  {
-    int left;
-    int right;
+    float left;
+    float right;
 }wheelSpeed_t;
 
 typedef struct  {
